some TypeErrors fixed

+++ src/Types/Chat.php
@@ -15,9 +15,9 @@ class Chat extends Base
     * Unique identifier for this chat. This number may be greater than 32 bits and some programming languages may have
     * difficulty/silent defects in interpreting it. But it is smaller than 52 bits, so a signed 64 bit integer or
     * double-precision float type are safe for storing this identifier.
-    * @var int
+    * @var float 
     */
-    public int $id;
+    public float $id;
 
     /**
     * Type of chat, can be either 'private', 'group', 'supergroup' or 'channel'
